Separate logic for Kafka and Yammer metrics
