+++ src/test/java/io/strimzi/kafka/metrics/http/ListenerTest.java
@@ -2,7 +2,7 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.kafka.metrics;
+package io.strimzi.kafka.metrics.http;
 
 import org.apache.kafka.common.config.ConfigException;
 import org.junit.jupiter.api.Test;
